Improve prepare_pdf.sh

+++ prepare_pdf.sh
@@ -3,6 +3,8 @@
 # Prepares pdf files for upload.
 
 for subject in matan algem; do
-  ./clean_pdf.sh $subject/$subject.pdf $subject.pdf
-  rm $subject/$subject.pdf
+  if [[ -e $subject/$subject.pdf ]] then
+    ./clean_pdf.sh $subject/$subject.pdf $subject.pdf
+    rm $subject/$subject.pdf
+  fi
 done
